+++ packages/evokit-plug/src/index.js
@@ -1,0 +1,5 @@
+import { createBlock } from 'evokit';
+
+export const Plug = createBlock('div', 'plug', [
+    'display',
+]);